Added course ownership(permission)

+++ controllers/courses.js
@@ -49,6 +49,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @Access Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId
+  req.body.user = req.user.id
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId)
 
@@ -57,6 +58,16 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
       new ErrorResponse(
         `Could not find bootcamp with the ID of: ${req.params.bootcampId}`,
         404
+      )
+    )
+  }
+
+  // make sure user is the bootcamp owner
+  if (req.user.id.toString() !== bootcamp.user && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
       )
     )
   }
@@ -84,6 +95,16 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
     )
   }
 
+  // make sure user is the course owner
+  if (req.user.id.toString() !== course.user && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    )
+  }
+
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
@@ -106,6 +127,17 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
       )
     )
   }
+
+  // make sure user is the course owner
+  if (req.user.id.toString() !== course.user && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        401
+      )
+    )
+  }
+
   course.deleteOne()
 
   res.status(200).json({
